# api-docs/swagger.yaml
basePath: /
definitions:
...
  handlers.JobSummary:
    properties:
      description:
        type: string
      name:
        type: string
      namespace:
        type: string
    type: object
...
  handlers.PackSummary:
    properties:
      description:
        type: string
      name:
        type: string
      namespace:
        type: string
    type: object
...
  handlers.RecentResource:
    properties:
      description:
        type: string
      id:
        type: integer
      latest_version:
        type: string
      name:
        type: string
      namespace:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
host: ramble.openwander.org
info:
  contact:
    name: API Support
  description: This is the API documentation for the RMBL Nomad Job & Pack Registry.
  license:
    name: MPL-2.0
  title: RMBL Nomad Registry API
  version: "1.0"
paths:
...
  /{username}/{resourcename}/raw:
    get:
      description: Get the raw .nomad.hcl content for the latest version of a resource.
      parameters:
      - description: User namespace
        in: path
        name: username
        required: true
        type: string
      - description: Resource name
        in: path
        name: resourcename
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Raw HCL content
          schema:
            type: string
      summary: Get raw HCL content
      tags:
      - resources
  /{username}/{resourcename}/v:
    get:
      description: Fetch metadata and README for a specific version of a resource.
        Usually called via HTMX.
      parameters:
      - description: User or Organization namespace
        in: path
        name: username
        required: true
        type: string
      - description: Resource name
        in: path
        name: resourcename
        required: true
        type: string
      - description: Version string (e.g., v1.0.0)
        in: query
        name: version
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragment
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get specific resource version details
      tags:
      - resources
  /{username}/{resourcename}/v/{version}/raw:
    get:
      description: Get the raw .nomad.hcl or template content for a specific version
        of a resource. Used by Nomad CLI.
      parameters:
      - description: User or Organization namespace
        in: path
        name: username
        required: true
        type: string
      - description: Resource name
        in: path
        name: resourcename
        required: true
        type: string
      - description: Version string
        in: path
        name: version
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Raw content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get raw HCL content for a specific version
      tags:
      - resources
...
  /docs/{page}:
    get:
      description: Renders a specific documentation page.
      parameters:
      - description: Doc page path
        in: path
        name: page
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML content
          schema:
            type: string
      summary: Get specific documentation page
      tags:
      - pages
  /jobs:
    get:
      description: Fetch a searchable and paginated list of all Nomad Jobs.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Sort order
        in: query
        name: sort
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML content
          schema:
            type: string
      summary: List nomad jobs
      tags:
      - resources
  /new:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Register a new Nomad job or pack in the registry.
      parameters:
      - description: Resource name
        in: formData
        name: name
        required: true
        type: string
      - description: Resource type (job, pack)
        in: formData
        name: type
        required: true
        type: string
      - description: Namespace owner (user or org:ID)
        in: formData
        name: owner
        required: true
        type: string
      - description: Git repository URL
        in: formData
        name: repository_url
        required: true
        type: string
      - description: Initial version
        in: formData
        name: version
        required: true
        type: string
      responses:
        "302":
          description: Redirect to new resource
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new resource
      tags:
      - resources
  /packs:
    get:
      description: Fetch a searchable and paginated list of all Nomad Packs.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Sort order
        in: query
        name: sort
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML content
          schema:
            type: string
      summary: List nomad packs
      tags:
      - resources
...
  /resource/{id}:
    delete:
      description: Permantently remove a resource and all its versions from the registry.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "403":
          description: Unauthorized
          schema:
            type: string
      summary: Delete a resource
      tags:
      - resources
  /resource/{id}/edit:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Update the metadata, repository URL, or tags for an existing resource.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: New resource name
        in: formData
        name: name
        required: true
        type: string
      - description: New resource type
        in: formData
        name: type
        required: true
        type: string
      - description: New description
        in: formData
        name: description
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "403":
          description: Unauthorized
          schema:
            type: string
      summary: Update resource details
      tags:
      - resources
  /resource/{id}/star:
    post:
      description: Star or unstar a resource for the authenticated user. Returns updated
        star button HTML.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragment
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Toggle star status
      tags:
      - resources
  /resource/{id}/webhook:
    post:
      consumes:
      - application/json
      description: Endpoint for receiving push and tag events from GitHub or GitLab.
        Automatically refreshes documentation or creates new versions.
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook secret
        in: query
        name: secret
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Receive Git webhooks
      tags:
      - webhooks
  /search:
    get:
      description: Search for jobs or packs with optional filtering by type and tag.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Resource type (job or pack)
        in: query
        name: type
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: HTML content
          schema:
            type: string
      summary: Search nomad resources
      tags:
      - resources
...
  /v1/jobs/search:
    get:
      description: Search for Nomad Jobs across all namespaces by name or description.
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handlers.JobSummary'
              type: array
            type: object
      summary: Search jobs
      tags:
      - nomad-job
...
  /v1/packs/search:
    get:
      description: Search for Nomad Packs across all namespaces by name or description.
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handlers.PackSummary'
              type: array
            type: object
      summary: Search packs
      tags:
      - nomad-pack
  /v1/recent:
    get:
      description: Fetch a list of recently updated packs and jobs, optionally filtered
        by time.
      parameters:
      - description: ISO8601 timestamp to filter resources updated after this time
        in: query
        name: since
        type: string
      - description: Maximum number of results (default 50, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/handlers.RecentResource'
              type: array
            type: object
      summary: List recently updated resources
      tags:
      - resources
...
  /v1/registries/search:
    get:
      description: Search for namespaces (users/orgs) that have published Nomad Packs.
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: Search registry namespaces
      tags:
      - nomad-pack
  /webhooks/github/issues:
    post:
      consumes:
      - application/json
      description: Webhook endpoint for GitHub to send issue events. Used to sync
        community pack/job requests with GitHub issues. Requires valid X-Hub-Signature-256
        header.
      parameters:
      - description: GitHub webhook signature
        in: header
        name: X-Hub-Signature-256
        required: true
        type: string
      - description: GitHub event type
        in: header
        name: X-GitHub-Event
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "403":
          description: Invalid signature
          schema:
            type: string
        "500":
          description: Webhook secret not configured
          schema:
            type: string
      summary: Receive GitHub issue webhooks
      tags:
      - webhooks
swagger: "2.0"